Simplify block view port code

+++ presto-main/src/test/java/com/facebook/presto/operator/aggregation/TestCountAggregation.java
@@ -11,7 +11,7 @@ public class TestCountAggregation
     extends AbstractTestAggregationFunction
 {
     @Override
-    public BlockCursor getSequenceCursor(long max)
+    public BlockCursor getSequenceCursor(int max)
     {
         return new LongSequenceCursor(max);
     }
